+++ CHANGES.rst
@@ -73,7 +73,7 @@ Improvements
 Bugfixes
 ^^^^^^^^
 
-- None so far :)
+- Fix editing service API calls using the service token (:pr:`5170`)
 
 
 Version 3.0.3